requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
anyio==4.9.0
attrs==25.3.0
beautifulsoup4==4.13.3
certifi==2025.1.31
charset-normalizer==3.4.1
frozenlist==1.5.0
greenlet==3.1.1
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.10
lxml==5.3.2
multidict==6.4.3
notion-client==2.3.0
propcache==0.3.1
pyee==12.1.1
python-dotenv==1.1.0
python-telegram-bot==20.0
requests==2.32.3
rfc3986==1.5.0
sniffio==1.3.1
soupsieve==2.6
typing_extensions==4.13.2
urllib3==2.4.0
yarl==1.19.0
